cleanup(core): clean up workspace creation
