+++ e2e/workspace-create/src/create-nx-workspace.test.ts
@@ -147,31 +147,6 @@ describe('create-nx-workspace', () => {
     expectNoAngularDevkit();
   });
 
-  it('should be able to create an angular + nest workspace', () => {
-    const wsName = uniq('angular-nest');
-    const appName = uniq('app');
-    runCreateWorkspace(wsName, {
-      preset: 'angular-nest',
-      style: 'css',
-      appName,
-      packageManager,
-    });
-  });
-
-  it('should be able to create an react + express workspace', () => {
-    const wsName = uniq('react-express');
-    const appName = uniq('app');
-    runCreateWorkspace(wsName, {
-      preset: 'react-express',
-      style: 'css',
-      appName,
-      packageManager,
-    });
-
-    expectNoAngularDevkit();
-    expectNoTsJestInJestConfig(appName);
-  });
-
   it('should be able to create an express workspace', () => {
     const wsName = uniq('express');
     const appName = uniq('app');
@@ -272,8 +247,6 @@ describe('create-nx-workspace', () => {
       cli: 'angular',
     });
 
-    const nxJson = readJson('nx.json');
-    expect(nxJson.cli.packageManager).toEqual('npm');
     checkFilesDoNotExist('yarn.lock');
     checkFilesExist('package-lock.json');
     process.env.SELECTED_PM = packageManager;
