+++ packages/angular/executors.json
@@ -15,26 +15,6 @@
       "schema": "./src/executors/package/schema.json",
       "description": "Builds and packages an Angular library to be distributed as an NPM package. It supports incremental builds."
     },
-    "webpack-browser": {
-      "implementation": "./src/builders/webpack-browser/webpack-browser.impl",
-      "schema": "./src/builders/webpack-browser/schema.json",
-      "description": "The `webpack-browser` executor is very similar to the standard `browser` builder provided by the Angular Devkit. It allows you to build your Angular application to a build artifact that can be hosted online. There are some key differences:   \n- Supports Custom Webpack Configurations  \n- Supports Incremental Building"
-    },
-    "webpack-dev-server": {
-      "implementation": "./src/builders/webpack-dev-server/webpack-dev-server.impl",
-      "schema": "./src/builders/webpack-dev-server/schema.json",
-      "description": "The `webpack-dev-server` executor is very similar to the standard `dev-server` builder provided by the Angular Devkit. It is usually used in tandem with `@nrwl/angular:webpack-browser` when your Angular application uses a custom webpack configuration."
-    },
-    "webpack-server": {
-      "implementation": "./src/builders/webpack-server/webpack-server.impl",
-      "schema": "./src/builders/webpack-server/schema.json",
-      "description": "The `webpack-server` executor is very similar to the standard `server` builder provided by the Angular Devkit. It is usually used in tandem with `@nrwl/angular:webpack-browser` when your Angular application uses a custom webpack configuration and NgUniversal for SSR."
-    },
-    "module-federation-dev-server": {
-      "implementation": "./src/builders/module-federation-dev-server/module-federation-dev-server.impl",
-      "schema": "./src/builders/module-federation-dev-server/schema.json",
-      "description": "The module-federation-dev-server executor is reserved exclusively for use with host Module Federation applications. It allows the user to specify which remote applications should be served with the host."
-    },
     "file-server": {
       "implementation": "./src/executors/file-server/file-server.impl",
       "schema": "./src/executors/file-server/schema.json",
