+++ packages/create-nx-workspace/bin/create-nx-workspace.ts
@@ -52,10 +52,8 @@ enum Preset {
   TS = 'ts',
   WebComponents = 'web-components',
   Angular = 'angular',
-  AngularWithNest = 'angular-nest',
   React = 'react',
   ReactExperimental = 'react-experimental',
-  ReactWithExpress = 'react-express',
   ReactNative = 'react-native',
   Expo = 'expo',
   NextJs = 'next',
@@ -110,16 +108,6 @@ const presetOptions: { name: Preset; message: string }[] = [
   {
     name: Preset.Expo,
     message: 'expo              [a workspace with a single Expo application]',
-  },
-  {
-    name: Preset.ReactWithExpress,
-    message:
-      'react-express     [a workspace with a full stack application (React + Express)]',
-  },
-  {
-    name: Preset.AngularWithNest,
-    message:
-      'angular-nest      [a workspace with a full stack application (Angular + Nest)]',
   },
 ];
 
@@ -625,8 +613,7 @@ async function determineCli(
   }
 
   switch (preset) {
-    case Preset.Angular:
-    case Preset.AngularWithNest: {
+    case Preset.Angular: {
       return Promise.resolve('angular');
     }
     default: {
@@ -668,7 +655,7 @@ async function determineStyle(
     },
   ];
 
-  if (![Preset.Angular, Preset.AngularWithNest].includes(preset)) {
+  if (![Preset.Angular].includes(preset)) {
     choices.push({
       name: 'styl',
       message: 'Stylus(.styl)     [ http://stylus-lang.com ]',
@@ -676,12 +663,7 @@ async function determineStyle(
   }
 
   if (
-    [
-      Preset.ReactWithExpress,
-      Preset.React,
-      Preset.ReactExperimental,
-      Preset.NextJs,
-    ].includes(preset)
+    [Preset.React, Preset.ReactExperimental, Preset.NextJs].includes(preset)
   ) {
     choices.push(
       {
@@ -868,7 +850,7 @@ async function createApp(
 
   const pmc = getPackageManagerCommand(packageManager);
 
-  const command = `new ${name} ${args} --collection=@nrwl/workspace/generators.json --cli=${cli}`;
+  const command = `new ${name} ${args}`;
 
   const workingDir = process.cwd().replace(/\\/g, '/');
   let nxWorkspaceRoot = `"${workingDir}"`;
@@ -1039,7 +1021,6 @@ function pointToTutorialAndCourse(preset: Preset) {
 
     case Preset.React:
     case Preset.ReactExperimental:
-    case Preset.ReactWithExpress:
     case Preset.NextJs:
       output.addVerticalSeparator();
       output.note({
@@ -1048,7 +1029,6 @@ function pointToTutorialAndCourse(preset: Preset) {
       });
       break;
     case Preset.Angular:
-    case Preset.AngularWithNest:
       output.addVerticalSeparator();
       output.note({
         title,
