+++ packages/workspace/project.json
@@ -33,17 +33,27 @@
           },
           {
             "input": "packages/workspace",
-            "glob": "**/files-npm/**",
+            "glob": "**/files-integrated-repo/**",
+            "output": "/"
+          },
+          {
+            "input": "packages/workspace",
+            "glob": "**/files-integrated-repo/**/.gitkeep",
+            "output": "/"
+          },
+          {
+            "input": "packages/workspace",
+            "glob": "**/files-package-based-repo/**",
+            "output": "/"
+          },
+          {
+            "glob": "**/files-packaged-based-repo/**/.gitkeep",
+            "input": "packages/workspace",
             "output": "/"
           },
           {
             "input": "packages/workspace",
             "glob": "**/files-root-app/**",
-            "output": "/"
-          },
-          {
-            "input": "packages/workspace",
-            "glob": "**/files-npm/**/.gitkeep",
             "output": "/"
           },
           {
