+++ packages/workspace/src/generators/new/files-integrated-repo/__dot__gitignore
