+++ packages/workspace/src/generators/new/files-integrated-repo/__dot__vscode/extensions.json__tmpl__
@@ -1,6 +1,6 @@
 {
   "recommendations": [
-    <% if(cli === 'angular') { %>
+    <% if(cliCommand === 'ng') { %>
     "angular.ng-template",<% }
     %>
     "nrwl.angular-console",
