+++ packages/workspace/src/generators/new/files-integrated-repo/package.json__tmpl__
