+++ packages/workspace/src/generators/new/files-integrated-repo/tools/tsconfig.tools.json
