+++ packages/workspace/src/generators/new/files-integrated-repo/tsconfig.base.json
