+++ packages/workspace/src/generators/new/files-package-based-repo/__dot__prettierignore
