+++ packages/workspace/src/generators/new/files-package-based-repo/package.json__tmpl__
