+++ packages/workspace/src/generators/new/files-root-app/__dot__prettierignore
