+++ packages/workspace/src/generators/new/files-root-app/tsconfig.base.json
