+++ packages/workspace/src/generators/new/new.ts
@@ -1,8 +1,6 @@
 import {
   addDependenciesToPackageJson,
-  convertNxGenerator,
   formatFiles,
-  getPackageManagerCommand,
   getWorkspacePath as devkitGetWorkspacePath,
   installPackagesTask,
   names,
@@ -11,26 +9,22 @@ import {
 } from '@nrwl/devkit';
 
 import { join } from 'path';
-import * as yargsParser from 'yargs-parser';
-import { spawn } from 'child_process';
-import { workspaceGenerator } from '../workspace/workspace';
-import { nxVersion } from '../../utils/versions';
 import { Preset } from '../utils/presets';
-import { getNpmPackageVersion } from '../utils/get-npm-package-version';
 import { Linter } from '../../utils/lint';
+import { generateWorkspaceFiles } from './generate-workspace-files';
+import { addPresetDependencies, generatePreset } from './generate-preset';
 
-export interface Schema {
-  cli: 'nx' | 'angular';
+interface Schema {
   directory: string;
   name: string;
-  appName: string;
+  appName?: string;
   npmScope?: string;
   skipInstall?: boolean;
   style?: string;
   nxCloud?: boolean;
   preset: string;
   defaultBase: string;
-  linter: Linter;
+  linter?: Linter;
   packageManager?: PackageManager;
 }
 
@@ -38,76 +32,28 @@ export interface NormalizedSchema extends Schema {
   presetVersion?: string;
 }
 
-function generatePreset(host: Tree, opts: NormalizedSchema) {
-  const cliCommand = opts.cli === 'angular' ? 'ng' : 'nx';
-  const parsedArgs = yargsParser(process.argv, {
-    boolean: ['interactive'],
-  });
-  const spawnOptions = {
-    stdio: [process.stdin, process.stdout, process.stderr],
-    shell: true,
-    cwd: join(host.root, opts.directory),
-  };
-  const pmc = getPackageManagerCommand();
-  const executable = `${pmc.exec} ${cliCommand}`;
-  const args = getPresetArgs(opts);
+export async function newGenerator(host: Tree, options: Schema) {
+  options = normalizeOptions(options);
+  validateOptions(options, host);
 
-  return new Promise<void>((resolve, reject) => {
-    spawn(executable, args, spawnOptions).on('close', (code: number) => {
-      if (code === 0) {
-        resolve();
-      } else {
-        const message = 'Workspace creation failed, see above.';
-        reject(new Error(message));
-      }
-    });
-  });
+  await generateWorkspaceFiles(host, { ...options, nxCloud: undefined } as any);
 
-  function getPresetArgs(options: NormalizedSchema) {
-    if (Object.values(Preset).some((val) => val === options.preset)) {
-      // supported presets
-      return getDefaultArgs(options);
+  addPresetDependencies(host, options);
+  addCloudDependencies(host, options);
+
+  await formatFiles(host);
+
+  return async () => {
+    installPackagesTask(host, false, options.directory, options.packageManager);
+    if (options.preset !== Preset.NPM && options.preset !== Preset.Core) {
+      await generatePreset(host, options);
     }
-    return getThirdPartyPresetArgs();
-  }
-
-  function getDefaultArgs(opts: Schema) {
-    return [
-      `g`,
-      `@nrwl/workspace:preset`,
-      `--name=${opts.appName}`,
-      opts.style ? `--style=${opts.style}` : null,
-      opts.linter ? `--linter=${opts.linter}` : null,
-      opts.npmScope ? `--npmScope=${opts.npmScope}` : `--npmScope=${opts.name}`,
-      opts.preset ? `--preset=${opts.preset}` : null,
-      opts.packageManager ? `--packageManager=${opts.packageManager}` : null,
-      `--cli=${cliCommand}`,
-      parsedArgs.interactive ? '--interactive=true' : '--interactive=false',
-    ].filter((e) => !!e);
-  }
-
-  function getThirdPartyPresetArgs() {
-    const thirdPartyPkgArgs = Object.entries(opts).reduce(
-      (acc, [key, value]) => {
-        if (value === true) {
-          acc.push(`--${key}`);
-        } else if (value === false) {
-          acc.push(`--no-${key}`);
-          // nxWorkspaceRoot breaks CLI if incorrectly set, so need to exclude it.
-          // TODO(jack): Should read in the preset schema and only pass the options specified.
-        } else if (key !== 'nxWorkspaceRoot') {
-          // string, number (don't handle arrays or nested objects)
-          acc.push(`--${key}=${value}`);
-        }
-        return acc;
-      },
-      []
-    );
-    return [`g`, `${opts.preset}:preset`, ...thirdPartyPkgArgs];
-  }
+  };
 }
 
-export async function newGenerator(host: Tree, options: Schema) {
+export default newGenerator;
+
+function validateOptions(options: Schema, host: Tree) {
   if (
     options.skipInstall &&
     options.preset !== Preset.Apps &&
@@ -128,8 +74,6 @@ export async function newGenerator(host: Tree, options: Schema) {
     );
   }
 
-  options = normalizeOptions(options);
-
   if (
     host.exists(options.name) &&
     !host.isFile(options.name) &&
@@ -139,110 +83,6 @@ export async function newGenerator(host: Tree, options: Schema) {
       `${join(host.root, options.name)} is not an empty directory.`
     );
   }
-
-  await workspaceGenerator(host, { ...options, nxCloud: undefined } as any);
-
-  addPresetDependencies(host, options);
-  addCloudDependencies(host, options);
-
-  await formatFiles(host);
-  return async () => {
-    installPackagesTask(host, false, options.directory, options.packageManager);
-    if (options.preset !== Preset.NPM && options.preset !== Preset.Core) {
-      await generatePreset(host, options);
-    }
-  };
-}
-
-export default newGenerator;
-export const newSchematic = convertNxGenerator(newGenerator);
-
-function addCloudDependencies(host: Tree, options: Schema) {
-  if (options.nxCloud) {
-    return addDependenciesToPackageJson(
-      host,
-      {},
-      { '@nrwl/nx-cloud': 'latest' },
-      join(options.directory, 'package.json')
-    );
-  }
-}
-
-function getPresetDependencies(preset: string, version?: string) {
-  switch (preset) {
-    case Preset.TS:
-      return { dependencies: {}, dev: { '@nrwl/js': nxVersion } };
-
-    case Preset.Angular:
-      return { dependencies: { '@nrwl/angular': nxVersion }, dev: {} };
-
-    case Preset.AngularWithNest:
-      return {
-        dependencies: { '@nrwl/angular': nxVersion },
-        dev: { '@nrwl/nest': nxVersion },
-      };
-
-    case Preset.Express:
-      return { dependencies: {}, dev: { '@nrwl/express': nxVersion } };
-
-    case Preset.Nest:
-      return { dependencies: {}, dev: { '@nrwl/nest': nxVersion } };
-
-    case Preset.NextJs:
-      return { dependencies: { '@nrwl/next': nxVersion }, dev: {} };
-
-    case Preset.React:
-      return { dependencies: {}, dev: { '@nrwl/react': nxVersion } };
-
-    case Preset.ReactExperimental:
-      return { dependencies: {}, dev: { '@nrwl/react': nxVersion } };
-
-    case Preset.ReactWithExpress:
-      return {
-        dependencies: {},
-        dev: {
-          '@nrwl/react': nxVersion,
-          '@nrwl/express': nxVersion,
-        },
-      };
-
-    case Preset.ReactNative:
-      return { dependencies: {}, dev: { '@nrwl/react-native': nxVersion } };
-
-    case Preset.Expo:
-      return { dependencies: {}, dev: { '@nrwl/expo': nxVersion } };
-
-    case Preset.WebComponents:
-      return { dependencies: {}, dev: { '@nrwl/web': nxVersion } };
-
-    default: {
-      return {
-        dev: {},
-        dependencies: { [preset]: version ?? getNpmPackageVersion(preset) },
-      };
-    }
-  }
-}
-
-function addPresetDependencies(host: Tree, options: NormalizedSchema) {
-  if (
-    options.preset === Preset.Apps ||
-    options.preset === Preset.Core ||
-    options.preset === Preset.Empty ||
-    options.preset === Preset.NPM
-  ) {
-    return;
-  }
-  const { dependencies, dev } = getPresetDependencies(
-    options.preset,
-    options.presetVersion
-  );
-  return addDependenciesToPackageJson(
-    host,
-    dependencies,
-    dev,
-    join(options.directory, 'package.json')
-  );
 }
 
 function normalizeOptions(options: NormalizedSchema): NormalizedSchema {
@@ -263,4 +103,15 @@ function normalizeOptions(options: NormalizedSchema): NormalizedSchema {
   }
 
   return options;
+}
+
+function addCloudDependencies(host: Tree, options: Schema) {
+  if (options.nxCloud) {
+    return addDependenciesToPackageJson(
+      host,
+      {},
+      { '@nrwl/nx-cloud': 'latest' },
+      join(options.directory, 'package.json')
+    );
+  }
 }
