+++ packages/workspace/src/generators/preset/preset.ts
@@ -1,24 +1,15 @@
 import {
   addDependenciesToPackageJson,
-  convertNxGenerator,
   formatFiles,
-  generateFiles,
   installPackagesTask,
   names,
   readWorkspaceConfiguration,
   Tree,
-  updateJson,
   updateWorkspaceConfiguration,
 } from '@nrwl/devkit';
-import { getImportPath } from 'nx/src/utils/path';
 import { Schema } from './schema';
-
-import { libraryGenerator } from '../library/library';
-
-import { insertImport } from '../utils/insert-import';
 import { insertStatement } from '../utils/insert-statement';
 import { Preset } from '../utils/presets';
-import { join } from 'path';
 
 export async function presetGenerator(tree: Tree, options: Schema) {
   options = normalizeOptions(options);
@@ -28,7 +19,6 @@ export async function presetGenerator(tree: Tree, options: Schema) {
     installPackagesTask(tree);
   };
 }
-export const presetSchematic = convertNxGenerator(presetGenerator);
 export default presetGenerator;
 
 async function createPreset(tree: Tree, options: Schema) {
@@ -101,60 +91,6 @@ async function createPreset(tree: Tree, options: Schema) {
       `apps/${names(options.name).fileName}/src/polyfills.ts`,
       ['@ungap/custom-elements']
     );
-  } else if (options.preset === Preset.AngularWithNest) {
-    const {
-      applicationGenerator: angularApplicationGenerator,
-    } = require('@nrwl' + '/angular/generators');
-    const { applicationGenerator: nestApplicationGenerator } = require('@nrwl' +
-      '/nest');
-
-    await angularApplicationGenerator(tree, {
-      name: options.name,
-      style: options.style,
-      linter: options.linter,
-      skipFormat: true,
-      standaloneConfig: options.standaloneConfig,
-    });
-    await nestApplicationGenerator(tree, {
-      name: 'api',
-      frontendProject: options.name,
-      linter: options.linter,
-      standaloneConfig: options.standaloneConfig,
-    });
-    await libraryGenerator(tree, {
-      name: 'api-interfaces',
-      unitTestRunner: 'none',
-      linter: options.linter,
-      standaloneConfig: options.standaloneConfig,
-    });
-    connectAngularAndNest(tree, options);
-  } else if (options.preset === Preset.ReactWithExpress) {
-    const {
-      applicationGenerator: expressApplicationGenerator,
-    } = require('@nrwl' + '/express');
-    const {
-      applicationGenerator: reactApplicationGenerator,
-    } = require('@nrwl' + '/react');
-
-    await reactApplicationGenerator(tree, {
-      name: options.name,
-      style: options.style,
-      linter: options.linter,
-      standaloneConfig: options.standaloneConfig,
-    });
-    await expressApplicationGenerator(tree, {
-      name: 'api',
-      frontendProject: options.name,
-      linter: options.linter,
-      standaloneConfig: options.standaloneConfig,
-    });
-    await libraryGenerator(tree, {
-      name: 'api-interfaces',
-      unitTestRunner: 'none',
-      linter: options.linter,
-      standaloneConfig: options.standaloneConfig,
-    });
-    connectReactAndExpress(tree, options);
   } else if (options.preset === Preset.Nest) {
     const { applicationGenerator: nestApplicationGenerator } = require('@nrwl' +
       '/nest');
@@ -199,202 +135,6 @@ async function createPreset(tree: Tree, options: Schema) {
   } else {
     throw new Error(`Invalid preset ${options.preset}`);
   }
-}
-
-function connectAngularAndNest(host: Tree, options: Schema) {
-  const { insertNgModuleImport } = require('@nrwl' +
-    '/angular/src/generators/utils');
-  host.write(
-    'libs/api-interfaces/src/lib/api-interfaces.ts',
-    `export interface Message { message: string }`
-  );
-
-  const modulePath = `apps/${options.name}/src/app/app.module.ts`;
-
-  insertImport(host, modulePath, 'HttpClientModule', '@angular/common/http');
-
-  insertNgModuleImport(host, modulePath, 'HttpClientModule');
-
-  const importScope = getImportPath(options.npmScope, '');
-  const scopePrefix = options.npmScope ? `${options.npmScope}-` : '';
-  const style = options.style ?? 'css';
-  host.write(
-    `apps/${options.name}/src/app/app.component.ts`,
-    `import { Component } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
-import { Message } from '${importScope}api-interfaces';
-
-@Component({
-  selector: '${scopePrefix}root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.${style}']
-})
-export class AppComponent {
-  hello$ = this.http.get<Message>('/api/hello')
-  constructor(private http: HttpClient) {}
-}
-    `
-  );
-
-  host.write(
-    `apps/${options.name}/src/app/app.component.spec.ts`,
-    `import { Component } from '@angular/core';
-import { TestBed, async } from '@angular/core/testing';
-import { HttpClientModule } from '@angular/common/http';
-import { AppComponent } from './app.component';
-
-describe('AppComponent', () => {
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [AppComponent],
-      imports: [HttpClientModule]
-    }).compileComponents();
-  }));
-
-  it('should create the app', () => {
-    const fixture = TestBed.createComponent(AppComponent);
-    const app = fixture.debugElement.componentInstance;
-    expect(app).toBeTruthy();
-  });
-});
-    `
-  );
-
-  host.write(
-    `apps/${options.name}/src/app/app.component.html`,
-    `<div style="text-align:center">
-  <h1>Welcome to ${options.name}!</h1>
-  <img
-    width="450"
-    src="https://raw.githubusercontent.com/nrwl/nx/master/images/nx-logo.png"
-    alt="Nx - Smart, Fast and Extensible Build System"
-  />
-</div>
-<div>Message: {{ (hello$|async)|json }}</div>
-    `
-  );
-
-  host.write(
-    `apps/api/src/app/app.controller.ts`,
-    `import { Controller, Get } from '@nestjs/common';
-
-import { Message } from '${importScope}api-interfaces';
-
-import { AppService } from './app.service';
-
-@Controller()
-export class AppController {
-  constructor(private readonly appService: AppService) {}
-
-  @Get('hello')
-  getData(): Message {
-    return this.appService.getData();
-  }
-}
-    `
-  );
-
-  host.write(
-    `apps/api/src/app/app.service.ts`,
-    `import { Injectable } from '@nestjs/common';
-import { Message } from '${importScope}api-interfaces';
-
-@Injectable()
-export class AppService {
-  getData(): Message {
-    return { message: 'Welcome to api!' };
-  }
-}
-    `
-  );
-}
-
-function connectReactAndExpress(host: Tree, options: Schema) {
-  const importScope = getImportPath(options.npmScope, '');
-  host.write(
-    'libs/api-interfaces/src/lib/api-interfaces.ts',
-    `export interface Message { message: string }`
-  );
-
-  host.write(
-    `apps/${options.name}/src/app/app.tsx`,
-    `import React, { useEffect, useState } from 'react';
-import { Message } from '${importScope}api-interfaces';
-
-export const App = () => {
-  const [m, setMessage] = useState<Message>({ message: '' });
-
-  useEffect(() => {
-    fetch('/api')
-      .then(r => r.json())
-      .then(setMessage);
-  }, []);
-
-  return (
-    <>
-      <div style={{ textAlign: 'center' }}>
-        <h1>Welcome to ${options.name}!</h1>
-        <img
-          width="450"
-          src="https://raw.githubusercontent.com/nrwl/nx/master/images/nx-logo.png"
-          alt="Nx - Smart, Fast and Extensible Build System"
-        />
-      </div>
-      <div>{m.message}</div>
-    </>
-  );
-};
-
-export default App;
-    `
-  );
-
-  host.write(
-    `apps/${options.name}/src/app/app.spec.tsx`,
-    `import { cleanup, getByText, render, waitFor } from '@testing-library/react';
-import React from 'react';
-import App from './app';
-
-describe('App', () => {
-  afterEach(() => {
-    delete global['fetch'];
-    cleanup();
-  });
-
-  it('should render successfully', async () => {
-    global['fetch'] = jest.fn().mockResolvedValueOnce({
-      json: () => ({
-        message: 'my message'
-      })
-    });
-
-    const { baseElement } = render(<App />);
-    await waitFor(() => getByText(baseElement, 'my message'));
-  });
-});
-    `
-  );
-
-  host.write(
-    `apps/api/src/main.ts`,
-    `import * as express from 'express';
-import { Message } from '${importScope}api-interfaces';
-
-const app = express();
-
-const greeting: Message = { message: 'Welcome to api!' };
-
-app.get('/api', (req, res) => {
-  res.send(greeting);
-});
-
-const port = process.env.port || 3333;
-const server = app.listen(port, () => {
-  console.log('Listening at http://localhost:' + port + '/api');
-});
-server.on('error', console.error);
-    `
-  );
 }
 
 function addPolyfills(host: Tree, polyfillsPath: string, polyfills: string[]) {
