+++ packages/workspace/src/generators/utils/presets.ts
@@ -6,10 +6,8 @@ export enum Preset {
   TS = 'ts',
   WebComponents = 'web-components',
   Angular = 'angular',
-  AngularWithNest = 'angular-nest',
   React = 'react',
   ReactExperimental = 'react-experimental',
-  ReactWithExpress = 'react-express',
   ReactNative = 'react-native',
   Expo = 'expo',
   NextJs = 'next',
