--- packages/workspace/src/generators/workspace/schema.d.ts
@@ -1,15 +1,0 @@
-import { PackageManager } from '@nrwl/devkit';
-
-export interface Schema {
-  name: string;
-  directory: string;
-  npmScope?: string;
-  skipInstall?: boolean;
-  skipGit?: boolean;
-  style?: string;
-  commit?: { name: string; email: string; message?: string };
-  cli: 'nx' | 'angular';
-  preset: string;
-  defaultBase: string;
-  packageManager?: PackageManager;
-}
