+++ packages/workspace/src/generators/workspace/schema.json
@@ -66,12 +66,6 @@
       "description": "The package manager used to install dependencies.",
       "type": "string",
       "enum": ["npm", "yarn", "pnpm"]
-    },
-    "cli": {
-      "description": "CLI used for generating code and running tasks",
-      "type": "string",
-      "enum": ["nx", "angular"],
-      "default": "nx"
     }
   }
 }
